requirements.txt
pycryptodome
sockets
cryptography